(SERVER-1803) Exclude org.ow2.asm deps from uberjar
Previously, jruby-core brought in discrete org.ow2.asm dependencies.
The versions of these dependencies conflict with the asm-all dependency
that Clojure's core.async uses.  When the jruby-deps uberjar were used
with other jars on the Java command line, the version of the asm
dependencies which would be used would depend on the order the jars
appear on the Java classpath.  In cases of conflicts between the two,
e.g., where a class might exist in one jar but not the other within the
same Java package, this could lead to problems where classes from
mismatched versions were errantly used together.

This commit changes the dependencies to export asm-all as an artifact
dependency rather than the discrete org.ow2.asm dependencies, to allow
for asm conflict resolution to be done at packaging time.  Also, all
of the asm dependencies are omitted from the uberjar profile so that
a jruby-deps uberjar doesn't run the risk of having asm content in it
which would conflict with the content in a paired jar used on the same
Java classpath.